+++ components/common/grid/observer.vue
@@ -1,0 +1,34 @@
+<template>
+  <div :data-test-id="`grid-${dataTestId}`" class="observer" />
+</template>
+
+<script>
+export default {
+  props: {
+    options: {
+      type: Object,
+      default: () => null,
+    },
+    dataTestId: {
+      type: String,
+      required: true,
+    },
+  },
+  data: () => ({
+    observer: null,
+  }),
+  mounted() {
+    const options = this.options || {}
+    this.observer = new IntersectionObserver(([entry]) => {
+      if (entry && entry.isIntersecting) {
+        this.$emit('intersect')
+      }
+    }, options)
+
+    this.observer.observe(this.$el)
+  },
+  destroyed() {
+    this.observer.disconnect()
+  },
+}
+</script>
